Implement major part of repository interfaces Minor code changes

+++ src/main/java/com/kirak/repository/AptTypeRepository.java
@@ -11,9 +11,9 @@ public interface AptTypeRepository {
 
     AptType save(AptType type);
 
-    boolean delete(int id);
+    boolean delete(short id);
 
-    AptType get(int id);
+    AptType get(short id);
 
     List<AptType> getAll();
 }
